:zap: 合并类
